# api/SWAGGER-API.yaml
openapi: 3.0.0
info:
  title: Docs Api Monopatines
  description: Utilizacion de api de uso y gestion de app de monopatines
  version: "2.0"
servers:
- url: https://virtserver.swaggerhub.com/LEOPOLDOVILLAO14/tpe-arquitecturas/2.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8080
  description: Swagger
...
paths:
...
  /usuarios/:
    get:
      tags:
      - Usuarios
      summary: Obtiene la lista de todos los usuarios-cliente.
      description: Ideal para listar todos los usuarios.
      responses:
        "200":
          description: Lista de usuarios obtenida con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    post:
      tags:
      - Usuarios
      summary: Crea un usuario asociado a una cuenta con rol cliente.
      description: Ideal para asociar varios usuarios a una cuenta.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserDTO'
        required: true
      responses:
        "201":
          description: Usuario creado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
  /usuarios/{id}:
    get:
      tags:
      - Usuarios
      summary: "Obtiene un usuario asociado a su {id}."
      description: Ideal para ver detalles de un usuario en partricular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id del usuario a buscar.
      responses:
        "200":
          description: Usuario obtenido con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
    put:
      tags:
      - Usuarios
      summary: "Modifica un usuario asociado a su {id}."
      description: Ideal para modificar propiedades de un usuario en partricular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id del usuario a modificar.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserVisibleDataDTO'
        required: true
      responses:
        "200":
          description: Usuario modificado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserVisibleDataDTO'
    delete:
      tags:
      - Usuarios
      summary: "Elimina un usuario asociado a su {id}."
      description: Ideal para eliminar un usuario en partricular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id del usuario a eliminar.
      responses:
        "200":
          description: Usuario eliminado con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserVisibleDataDTO'
...
components:
  schemas:
...
    Usuario:
      type: object
      properties:
        id_user:
          type: number
          example: 25
        name:
          type: string
          example: juan
        surname:
          type: string
          example: hernandez
        phone:
          type: string
          example: "542494332211"
        email:
          type: string
          example: juanfer@gmail.com
        registrationDate:
          type: string
          example: 2024-10-05
        id_account:
          type: number
          example: 45
    UserVisibleDataDTO:
      required:
      - email
      - name
      - phone
      - surname
      type: object
      properties:
        name:
          type: string
          example: franco
        surname:
          type: string
          example: lopez
        phone:
          type: string
          example: "2494885533"
        email:
          type: string
          example: flopez@gmail.com
    NewUserDTO:
      required:
      - email
      - id_account
      - name
      - phone
      - surname
      type: object
      properties:
        name:
          type: string
          example: juan
        surname:
          type: string
          example: hernandez
        phone:
          type: string
          example: "2494009988"
        email:
          type: string
          example: juanfer@gmail.com
        id_account:
          type: number
          example: 45
    UserDTO:
      type: object
      properties:
        name:
          type: string
          example: juan
        surname:
          type: string
          example: hernandez
        phone:
          type: string
          example: "2494009988"
        email:
          type: string
          example: juanfer@gmail.com
        registrationDate:
          type: string
          example: 2024-11-03
...